minor changes, refactoring, adding readme

+++ src/main.cpp
@@ -10,10 +10,9 @@ int main()
 
     int constant = addConstant(chunk, 1.2);
     writeChunk(chunk, OP_CONSTANT, 123);
-    writeChunk(chunk, constant, 123);
+    writeChunk(chunk, static_cast<uint8_t>(constant), 123);
 
     constant = addConstant(chunk, 3.4);
-
     writeChunk(chunk, OP_CONSTANT, 123);
     writeChunk(chunk, static_cast<std::uint8_t>(constant), 123);
